sim: Move ProbeListenerObject to its own file
This removes the cyclic dependency between the sim
objects and probes files.
