+++ src/sim/probe/probe_listener_object.hh
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2013 ARM Limited
+ * Copyright (c) 2022-2023 The University of Edinburgh
  * All rights reserved
  *
  * The license below extends only to copyright in the software and shall
@@ -38,84 +39,40 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "sim/probe/probe.hh"
+#ifndef __SIM_PROBE_PROBE_LISTENER_OBJECT_HH__
+#define __SIM_PROBE_PROBE_LISTENER_OBJECT_HH__
 
-#include "debug/ProbeVerbose.hh"
+#include <vector>
+
+#include "sim/probe/probe.hh"
+#include "sim/sim_object.hh"
 
 namespace gem5
 {
 
-ProbePoint::ProbePoint(ProbeManager *manager, const std::string& _name)
-    : name(_name)
-{
-    if (manager) {
-        manager->addPoint(*this);
-    }
-}
+struct ProbeListenerObjectParams;
 
-ProbeListener::ProbeListener(ProbeManager *_manager, const std::string &_name)
-    : manager(_manager), name(_name)
+/**
+ * This class is a minimal wrapper around SimObject. It is used to declare
+ * a python derived object that can be added as a ProbeListener to any other
+ * SimObject.
+ *
+ * It instantiates manager from a call to Parent.any.
+ * The vector of listeners is used simply to hold onto listeners until the
+ * ProbeListenerObject is destroyed.
+ */
+class ProbeListenerObject : public SimObject
 {
-    manager->addListener(name, *this);
-}
+  protected:
+    ProbeManager *manager;
+    std::vector<ProbeListener *> listeners;
 
-ProbeListener::~ProbeListener()
-{
-    manager->removeListener(name, *this);
-}
-
-bool
-ProbeManager::addListener(std::string point_name, ProbeListener &listener)
-{
-    DPRINTFR(ProbeVerbose, "Probes: Call to addListener to \"%s\" on %s.\n",
-        point_name, name());
-    bool added = false;
-    for (auto p = points.begin(); p != points.end(); ++p) {
-        if ((*p)->getName() == point_name) {
-            (*p)->addListener(&listener);
-            added = true;
-        }
-    }
-    if (!added) {
-        DPRINTFR(ProbeVerbose, "Probes: Call to addListener to \"%s\" on "
-            "%s failed, no such point.\n", point_name, name());
-    }
-    return added;
-}
-
-bool
-ProbeManager::removeListener(std::string point_name, ProbeListener &listener)
-{
-    DPRINTFR(ProbeVerbose, "Probes: Call to removeListener from \"%s\" on "
-        "%s.\n", point_name, name());
-    bool removed = false;
-    for (auto p = points.begin(); p != points.end(); ++p) {
-        if ((*p)->getName() == point_name) {
-            (*p)->removeListener(&listener);
-            removed = true;
-        }
-    }
-    if (!removed) {
-        DPRINTFR(ProbeVerbose, "Probes: Call to removeListener from \"%s\" "
-            "on %s failed, no such point.\n", point_name, name());
-    }
-    return removed;
-}
-
-void
-ProbeManager::addPoint(ProbePoint &point)
-{
-    DPRINTFR(ProbeVerbose, "Probes: Call to addPoint \"%s\" to %s.\n",
-        point.getName(), name());
-
-    for (auto p = points.begin(); p != points.end(); ++p) {
-        if ((*p)->getName() == point.getName()) {
-            DPRINTFR(ProbeVerbose, "Probes: Call to addPoint \"%s\" to %s "
-                "failed, already added.\n", point.getName(), name());
-            return;
-        }
-    }
-    points.push_back(&point);
-}
+  public:
+    ProbeListenerObject(const ProbeListenerObjectParams &params);
+    virtual ~ProbeListenerObject();
+    ProbeManager* getProbeManager() { return manager; }
+};
 
 } // namespace gem5
+
+#endif//__SIM_PROBE_PROBE_LISTENER_OBJECT_HH__
